requirements.txt
asgiref==3.8.1
autopep8==2.3.2
black==25.1.0
chardet==5.2.0
click==8.1.8
cssselect2==0.7.0
Django==5.1.6
django-debug-toolbar==5.0.1
django-filer==3.3.1
django-polymorphic==3.1.0
django-silk==5.3.2
djangorestframework==3.15.2
easy-thumbnails==2.10
gprof2dot==2024.6.6
lxml==5.3.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pycodestyle==2.12.1
reportlab==4.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
webencodings==0.5.1
